Cleanup System.Environment implementation
The full netstandard20 Environment implementation lives in corefx. This change is first pass on getting things cleanup - with the eventual goal of not having public Environment type in CoreLib at all.
