+++ src/Common/src/Interop/Unix/System.Private.CoreLib.Native/Interop.Environment.cs
@@ -13,9 +13,6 @@ internal static partial class Interop
         [DllImport(Interop.Libraries.CoreLibNative, EntryPoint = "CoreLibNative_GetEnvironmentVariable")]
         internal static unsafe extern int GetEnvironmentVariable(string name, out IntPtr result);
 
-        [DllImport(Interop.Libraries.CoreLibNative, EntryPoint = "CoreLibNative_GetMachineName")]
-        internal static unsafe extern int GetMachineName(byte *hostNameBuffer, int hostNameBufferSize);
-
         [DllImport(Interop.Libraries.CoreLibNative, EntryPoint = "CoreLibNative_ExitProcess")]
         internal static extern void ExitProcess(int exitCode);
     }
